+++ packages/stream_feeds/test/test_utils/event_types.dart
@@ -1,0 +1,10 @@
+class EventTypes {
+  static const String followCreated = 'feeds.follow.created';
+  static const String followDeleted = 'feeds.follow.deleted';
+  static const String followUpdated = 'feeds.follow.updated';
+
+  static const String pollClosed = 'feeds.poll.closed';
+  static const String pollDeleted = 'feeds.poll.deleted';
+  static const String pollVoteCasted = 'feeds.poll.vote_casted';
+  static const String pollVoteRemoved = 'feeds.poll.vote_removed';
+}