+++ packages/stream_feeds/CHANGELOG.md
@@ -1,3 +1,6 @@
+## unreleased
+- Update follower and following counts on the feed state when receiving follow websocket events.
+
 ## 0.3.1
 - Update API client with renaming `addReaction` to `addActivityReaction` and `deleteReaction` to `deleteActivityReaction`.
 - Update `activity.currentFeed` capabilities when adding or updating activity from websocket events.